refactor: split useAuth context and provider in AuthProvider, AuthContext and useAuth
